backend/requirements.txt
# Core API
flask==3.0.3
flask-cors==4.0.0
PyJWT==2.8.0

# Test framework
pytest==8.2.0
pytest-cov==5.0.0
pytest-html==4.1.1

# Integration / E2E
requests==2.32.3
playwright==1.44.0

# Performance testing
# gevent pinned to 24.x which ships pre-built wheels for Python 3.12+
# Install with: pip install --prefer-binary -r requirements.txt
gevent==24.2.1
locust==2.28.0

# Security static analysis
bandit==1.7.9
